services/importer/requirements.txt
python-bitcoinrpc==1.0
neo4j==5.15.0
python-dotenv==1.0.0
tqdm==4.66.1
